feat - Adiciona rate limiting e validação de senha na deleção
Foram implementadas políticas de rate limiting para endpoints sensíveis (login, cadastro PF/PJ e troca de senha) usando o middleware do ASP.NET Core. Criadas classes específicas para cada política e extensão para registro centralizado. Program.cs ajustado para ativar o rate limiting. Endpoints protegidos com [EnableRateLimiting]. Agora, a exclusão de usuário exige confirmação de senha via novo DTO DeleteAccountDto. Refatorações menores e melhorias na organização dos arquivos.

+++ BlazorShop.Api/Services/Usuarios/UsuarioService.cs
@@ -3,6 +3,7 @@ using BlazorShop.Api.Repositories.Interfaces;
 using BlazorShop.Api.Security.Password.Intefaces;
 using BlazorShop.Api.Services.Usuarios.Interfaces;
 using BlazorShop.Models.Commons;
+using BlazorShop.Models.DTOs.UsuarioDtos.Delete;
 using BlazorShop.Models.DTOs.UsuarioDtos.Update;
 
 namespace BlazorShop.Api.Services.Usuarios;
@@ -59,13 +60,15 @@ public class UsuarioService(IPasswordService passwordService,
         return OperationResult<Usuario>.Ok(user);
     }
 
-    public async Task<OperationResult<Usuario>> DeleteUsuarioAsync(int id)
+    public async Task<OperationResult<Usuario>> DeleteUsuarioAsync(int id, DeleteAccountDto deleteAccountDto)
     {
         var usuario = await _usuarioRepository.GetByIdAsync(id);
         if (usuario is null)
-        {
-            return OperationResult<Usuario>.Fail("Usuario Não encontrado para deletar");
-        }
+            return OperationResult<Usuario>.Fail("Usuario Não encontrado!");
+        if (!_passwordService.IsInvalidPassword(usuario
+            , hashedPassword: usuario.Senha
+            , password: deleteAccountDto.Senha))
+            return OperationResult<Usuario>.Fail("Senha Incorreta");
         await _usuarioRepository.DeleteUsuarioAsync(usuario);
         return OperationResult<Usuario>.Ok(usuario, message: "Usuario deletado com sucesso");
